Fix db id issues + new dataset entity

+++ Geonorge.Forvaltningsinformasjon.Infrastructure/Persistence/Entities/NativeEntities/Fdvdatamottak.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public int? FdvrundeId { get; set; }
         public int? FdvpartId { get; set; }
-        public long? FdvdatasettId { get; set; }
+        public int? FdvdatasettId { get; set; }
         public string Dato { get; set; }
         public string Brukernavn { get; set; }
         public string Kommentar { get; set; }
